requirements.txt
coveralls==1.3.0
Django==2.0.5
django-filter==2.0.0
django-rest-framework==0.1.0
djangorestframework==3.8.2
gunicorn==19.8.1
Markdown==2.6.11
Pillow==5.1.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
pytest==3.6.4
pytz==2018.4
PyYAML==3.13
